pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-utils"
version = "0.1.0"
description = "Media utilities: Image generation with Z-Image-Turbo and LLM with Qwen"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["image-generation", "z-image", "qwen", "llm", "diffusion", "ai", "media"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "diffusers @ git+https://github.com/huggingface/diffusers",
    "transformers>=4.40.0",
    "huggingface_hub>=0.20.0",
    "accelerate>=0.25.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[project.urls]
Homepage = "https://github.com/slahiri/media-utils"
Repository = "https://github.com/slahiri/media-utils"
Changelog = "https://github.com/slahiri/media-utils/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["media_utils*"]

[tool.setuptools.package-data]
media_utils = ["../config.yaml"]
